refactor(sample): remove PreflightAction typealias
